[#49] apt-update fails when building environment with vagrant
Had to rollback vagrant box from ubuntu 13.10 to 12.04 because saucy is not longer supported by apt.puppetlabs.com.

This caused a few things to break, which had to be fixed:
* facter fqdn wasn't picking up epcc.local, so I had to add epcc to the node list
* apache version downgraded from 2.4 to 2.2 and was incompatible with our config file

Also disabled simulation of low-resource environment by default, so that non-dev users who want to run the app locally can do so with better performance.