Switch to using tooltips for help buttons
The tooltips are loaded at page creation, instead of using ajax, so they display instantly.
They're much more compact, though styling still needs work.

With this change, the sliding window from the top is *depreciated*, and will be removed in a future update.
It's ugly and the way it changed the layout of the page was annoying.