PHP include jquery, jquery-ui, and normalize.css
This minorly increases server load/bandwidth, but means the client doesn't have to make multiple requests.

Some of the files had to be modified to work with PHP includes.
Notes for updating are in the appropriate folders.

At this point the only things downloaded from off server are backgrounds and fonts.