Added generated ID to compare EPAtom objects.
In EPAtom.php added a private variable named TypeID, this variable is assigned an MD5 hash at the constructor and in the function loadSavePack,
this hash is generated from the name variable. Also added the function getTypeID and replaced the function getUid in the function match.

It's now possible for the character creator to load EPAtom classes/its children from other sessions and display the data correctly,
this is also backwards compatible with older characters since nothing is changed to the save format.