docker support
This change allows you to run EPCC in a Docker container. Docker Compose
is used to automate all of the necessary steps, including running MySQL
in a separate container. MySQL data is stored in a Docker volume, so the
data persists even if you tear down the containers. The source code is
mounted directly in the application container to support live editing.

I moved the FullDatabase.sql file to its own directory (src/sql/init) so
that the mysql docker container could source it without sourcing the
other SQL files. I wish the docker image supported filters, or the
ability to select a specific file, but alas that is not the case.

The EPConfigFile class has been modified to support an override of the
config file path via environment variable EPCC_CONFIG_PATH. This was
needed so that the epcc docker container can use a different database
connection URI, and support live editing. Without this, the container
would have to make a copy of the source code and change the config file.